test of unlockAll, overhaul of Apostrophe tests folder so you can just do mocha test/whateverfile.js
