+++ test/attachments.js
@@ -1,14 +1,14 @@
-
+var t = require('../test-lib/test.js');
 var assert = require('assert');
 var _ = require('lodash');
-;
+
 
 var apos;
 
 describe('Attachment', function() {
 
   after(function(done) {
-    return destroy(apos, done);
+    return t.destroy(apos, done);
   });
 
   this.timeout(5000);
