+++ test/db.js
@@ -1,3 +1,4 @@
+var t = require('../test-lib/test.js');
 var assert = require('assert');
 
 var apos;
@@ -5,7 +6,7 @@ var apos;
 describe('Db', function(){
 
   after(function(done) {
-    return destroy(apos, done);
+    return t.destroy(apos, done);
   });
 
   this.timeout(5000);
