+++ test/images.js
@@ -1,8 +1,8 @@
 
+var t = require('../test-lib/test.js');
 var assert = require('assert');
 var _ = require('lodash');
 var async = require('async');
-;
 
 var apos;
 
@@ -41,7 +41,7 @@ describe('Images', function() {
   this.timeout(5000);
 
   after(function(done) {
-    return destroy(apos, done);
+    return t.destroy(apos, done);
   });
 
   it('should be a property of the apos object', function(done) {
