+++ test/launder.js
@@ -1,3 +1,4 @@
+var t = require('../test-lib/test.js');
 var assert = require('assert');
 
 describe('Launder', function(){
@@ -5,7 +6,7 @@ describe('Launder', function(){
   this.timeout(5000);
 
   after(function(done) {
-    return destroy(apos, done);
+    return t.destroy(apos, done);
   });
 
   var apos;
