+++ test/locks.js
@@ -1,7 +1,7 @@
+var t = require('../test-lib/test.js');
 var assert = require('assert');
 var _ = require('lodash');
 var async = require('async');
-;
 
 var apos;
 
@@ -10,7 +10,7 @@ describe('Locks', function() {
   this.timeout(5000);
 
   after(function(done) {
-    return destroy(apos, done);
+    return t.destroy(apos, done);
   });
 
   it('should be a property of the apos object', function(done) {
