+++ test/oembed.js
@@ -1,4 +1,5 @@
 
+var t = require('../test-lib/test.js');
 var assert = require('assert');
 var _ = require('lodash');
 var async = require('async');
@@ -11,7 +12,7 @@ describe('Oembed', function() {
   this.timeout(5000);
 
   after(function(done) {
-    return destroy(apos, done);
+    return t.destroy(apos, done);
   });
 
   //////
